refactor(interface): dashboard topbar chrome (drop nav, gear→settings overlay) + mockup activity labels
