+++ apps/armada-interface/src/components/tx/stageCopy.test.ts
@@ -1,5 +1,5 @@
 // ABOUTME: Unit tests for stageCopy / kindTitle / recordTitle.
-// ABOUTME: Covers active/waiting executionState resolution (shield) and chain-name appending for xchain kinds.
+// ABOUTME: Covers active/waiting executionState resolution (shield) and the mockup activity copy (deposit source chain, sent-to-recipient, vault ops).
 
 import { describe, it, expect } from 'vitest'
 import { stageCopy, kindTitle, recordTitle } from './stageCopy'
@@ -41,7 +41,7 @@ describe('kindTitle', () => {
 })
 
 describe('recordTitle', () => {
-  it('returns the bare kind title for non-xchain kinds', () => {
+  it('names the source chain for a deposit (shield)', () => {
     const record: TxRecord<'shield'> = {
       id: '01J', kind: 'shield', executionState: 'pending', stage: 'build-proof',
       stagesCompleted: [], updatedSeq: 0, createdAt: 0, updatedAt: 0,
@@ -49,10 +49,10 @@ describe('recordTitle', () => {
       artifacts: {},
       walletContext: { evmAddress: undefined, shieldedWalletId: '', sourceChainId: 31337 },
     }
-    expect(recordTitle(record)).toBe('Deposit')
+    expect(recordTitle(record)).toMatch(/^Deposit from /)
   })
 
-  it('appends chain name for unshield-xchain', () => {
+  it('names the 0x recipient for a public unshield (external send / withdraw)', () => {
     const record: TxRecord<'unshield-xchain'> = {
       id: '01J', kind: 'unshield-xchain', executionState: 'pending', stage: 'build-proof',
       stagesCompleted: [], updatedSeq: 0, createdAt: 0, updatedAt: 0,
@@ -60,14 +60,60 @@ describe('recordTitle', () => {
         amount: 0n,
         feeCacheId: '',
         toChainId: 31338,
-        recipient: '0x0',
+        recipient: '0x1234567890abcdef1234567890abcdef12345678',
         broadcasterFeeAmount: 0n,
         broadcasterShieldedAddress: '',
       },
       artifacts: {},
       walletContext: { evmAddress: undefined, shieldedWalletId: '', sourceChainId: 31337 },
     }
-    expect(recordTitle(record)).toMatch(/^Withdraw to /)
+    expect(recordTitle(record)).toMatch(/^Sent to 0x/)
   })
 
+  it('reads "Sent to private address" for a private (0zk) transfer', () => {
+    const record: TxRecord<'transfer-shielded'> = {
+      id: '01J', kind: 'transfer-shielded', executionState: 'pending', stage: 'build-proof',
+      stagesCompleted: [], updatedSeq: 0, createdAt: 0, updatedAt: 0,
+      meta: {
+        amount: 0n,
+        feeCacheId: '',
+        recipient: '0zkaaaa',
+        broadcasterFeeAmount: 0n,
+        broadcasterShieldedAddress: '',
+      },
+      artifacts: {},
+      walletContext: { evmAddress: undefined, shieldedWalletId: '', sourceChainId: 31337 },
+    }
+    expect(recordTitle(record)).toBe('Sent to private address')
+  })
+
+  it('reads "Received payment" for an incoming private transfer', () => {
+    const record: TxRecord<'transfer-shielded-received'> = {
+      id: '01J', kind: 'transfer-shielded-received', executionState: 'completed', stage: 'observed',
+      stagesCompleted: ['observed'], updatedSeq: 0, createdAt: 0, updatedAt: 0,
+      meta: { amount: 0n },
+      artifacts: {},
+      walletContext: { evmAddress: undefined, shieldedWalletId: '', sourceChainId: 31337 },
+    }
+    expect(recordTitle(record)).toBe('Received payment')
+  })
+
+  it('reads the earn-vault copy for yield ops', () => {
+    const base = {
+      id: '01J', executionState: 'pending' as const, stage: 'build-proof',
+      stagesCompleted: [], updatedSeq: 0, createdAt: 0, updatedAt: 0,
+      artifacts: {},
+      walletContext: { evmAddress: undefined, shieldedWalletId: '', sourceChainId: 31337 },
+    }
+    const deposit: TxRecord<'yield-deposit'> = {
+      ...base, kind: 'yield-deposit',
+      meta: { amount: 0n, feeCacheId: '', broadcasterFeeAmount: 0n, broadcasterShieldedAddress: '' },
+    }
+    const withdraw: TxRecord<'yield-withdraw'> = {
+      ...base, kind: 'yield-withdraw',
+      meta: { amount: 0n, feeCacheId: '', shares: 0n, broadcasterFeeAmount: 0n, broadcasterShieldedAddress: '' },
+    }
+    expect(recordTitle(deposit)).toBe('Added to earn vault')
+    expect(recordTitle(withdraw)).toBe('Withdrawn from earn vault')
+  })
 })
