+++ apps/armada-interface/src/main.tsx
@@ -37,7 +37,6 @@ setTheme(getSavedTheme() ?? 'light')
 
 import { Dashboard } from '@/pages/Dashboard'
 import { History } from '@/pages/History'
-import { Settings } from '@/pages/Settings'
 import { AddressBook } from '@/pages/AddressBook'
 import { Debug } from '@/pages/Debug'
 
@@ -86,7 +85,6 @@ createRoot(document.getElementById('root')!).render(
               <Route element={<App />}>
                 <Route index element={<Dashboard />} />
                 <Route path="history" element={<History />} />
-                <Route path="settings" element={<Settings />} />
                 <Route path="address-book" element={<AddressBook />} />
                 {/* Debug page is available in both modes — contract addresses + per-chain
                     balances are useful diagnostics regardless. The local-only faucet UI is
