+++ apps/armada-interface/src/state/ui.ts
@@ -11,13 +11,14 @@ export type ModalKind =
   | 'yield-withdraw'
   | 'payment'
   | 'receive'
+  | 'settings'
   | 'wallet-unlock'
   | 'wallet-reset'
 
 /** Dashboard / action flows that require a connected EVM wallet before opening. */
 export type ActionModalKind = Exclude<
   ModalKind,
-  null | 'wallet-unlock' | 'wallet-reset' | 'receive'
+  null | 'wallet-unlock' | 'wallet-reset' | 'receive' | 'settings'
 >
 
 export const openModalAtom = atom<ModalKind>(null)
